Added more to checkmark mode

+++ main.py
@@ -11,35 +11,46 @@ txtPath = './lyricGeneration/text/'
 #skipYT = True
 skipYT = False
 
+#justTrain = True
+justTrain = False
+
+#justGenerate = True
+justGenerate = False
+
 checkmark = True
 #checkmark = False
 
-epoch = 1000
+#epoch = 1000
+epoch = 1
 
 # bin setup
 
-os.system('rm -r bin')
-os.makedirs('./bin')
-os.makedirs('./bin/model')
-os.makedirs('./bin/output')
+if not justGenerate or not justTrain:
+	os.system('rm -r bin')
+	os.makedirs('./bin')
+	os.makedirs('./bin/model')
+	os.makedirs('./bin/output')
 
 
 
-if not skipYT:
+if not (skipYT and not justGenerate) and not justTrain:
 	url = sys.argv[1]
 
 	print 'Converting Youtube video to MIDI. . .'
 	ytc.YoutubeToMIDIConvert(url)
 
+
 ## Use lyric generator
 lg.lyricGenerator(txtPath)
 
-# Use song composer
-print 'Generating numpy matrix of song'
-mg.matrixGenerate()
+if not justGenerate:
 
-print 'Training. . .'
-train.train(epoch, checkmark)
+	# Use song composer
+	print 'Generating numpy matrix of song'
+	mg.matrixGenerate()
+
+	print 'Training. . .'
+	train.train(epoch, checkmark)
 
 print 'Generating midi file. . .'
 gen.generate()
